+++ python/tank_vendor/shotgun_deploy/__init__.py
@@ -12,4 +12,4 @@ from .descriptor import Descriptor, create_descriptor
 from .io_descriptor import descriptor_dict_to_uri, descriptor_uri_to_dict
 from .manager import ToolkitManager
 from .paths import get_bundle_cache_root
-from .reload import swap_core
+from .import_handler import swap_core
