+++ python/tank_vendor/shotgun_deploy/import_handler.py
@@ -23,8 +23,7 @@ class CoreImportHandler(object):
         >>> # import/run a bunch of code
         >>> # context change, need to use a different core
         >>> importer.set_core_path("/path/to/a/different/version/of/tk-core")
-        >>> # namespaces cleared out and re-imported from new core location
-        >>> # new imports will come from new core location
+        >>> # namespaces cleared. new imports will come from new core location
 
     When an instance of this object is added to `sys.meta_path`, it is used to
     alter the way python imports packages.
@@ -45,7 +44,7 @@ class CoreImportHandler(object):
 
         """
 
-        self._core_path = None   # will be set shortly
+        self._core_path = None
         self._namespaces = []
 
         # a dictionary to hold module information after it is found, before
@@ -56,6 +55,12 @@ class CoreImportHandler(object):
         self.set_core_path(core_path)
 
     def __repr__(self):
+        """
+        A unique representation of the handler.
+
+        :return: str representation.
+        """
+
         return (
             "<CoreImportHandler for core located in: '%s'" % (self.core_path,))
 
@@ -75,33 +80,36 @@ class CoreImportHandler(object):
         :returns: this object (also a loader) if module found, None otherwise.
         """
 
-        # get the package name from (first part of the module fullname)
+        # get the package name (first part of the module fullname)
         module_path_parts = module_fullname.split('.')
         package_name = module_path_parts[0]
 
-        # make sure the package is in the list of supplied namespaces before
-        # continuing.
+        # make sure the package is in the list of namespaces before continuing.
         if package_name not in self._namespaces:
-            # the package is not in one of the supplied namespaces, returning
+            # the package is not in one of the core namespaces. returning
             # None tells python to use the next importer available (likely the
             # default import mechanism).
             return None
 
-        # the name of the module (without the module path)
+        # module path without the target module name
         module_name = module_path_parts.pop()
 
-        # the module path as an actual partial path on disk
+        # partial path to the module on disk
         if len(module_path_parts):
             module_path = os.path.join(*module_path_parts)
         else:
             module_path = ""
 
+        # if it exists, this should be the full directory path to the module in
+        # the current core.
         path = os.path.join(self.core_path, module_path)
 
         try:
             # find the module and store its info in a lookup based on the
             # full module name. The module info is a tuple of the form:
+            #
             #   (file_obj, filename, description)
+            #
             # If this find is successful, we'll need the info in order
             # to load it later.
             module_info = imp.find_module(module_name, [path])
@@ -132,17 +140,17 @@ class CoreImportHandler(object):
             # retrieve the found module info
             (file_obj, filename, desc) = self._module_info[module_fullname]
 
-            # attempt to load the module given the info from find_module
-            module = sys.modules.setdefault(
-                module_fullname,
-                imp.load_module(module_fullname, file_obj, filename, desc)
-            )
+            # attempt to load the module. if this fails, allow it to raise
+            # the usual `ImportError`
+            module = imp.load_module(module_fullname, file_obj, filename, desc)
         finally:
             # as noted in the imp.load_module docs, must close the file handle.
             if file_obj:
                 file_obj.close()
 
-            # no need to carry around the module info now that we've loaded it
+            # no need to carry around the module info now that we've loaded it.
+            # once the module is in `sys.modules`, the custom importer will
+            # no longer run.
             del self._module_info[module_fullname]
 
         # the module needs to know the loader so that reload() works
@@ -156,8 +164,8 @@ class CoreImportHandler(object):
 
         This method clears out `sys.modules` of all previously imported modules
         and reimports them using the new core path. This method locks the global
-        interpreter in an attempt to prevent problems from modifying sys.modules
-        in a multithreaded context.
+        interpreter to prevent problems from modifying sys.modules in a
+        multithreaded context.
 
         :param path: str path to the core to import from.
 
@@ -191,10 +199,6 @@ class CoreImportHandler(object):
         # same time.
         imp.acquire_lock()
 
-        # keep a runnng list of modules that need to be re-imported from the new
-        # core location.
-        module_names_to_import = []
-
         # sort by package depth, deeper modules first
         module_names = sorted(
             sys.modules.keys(),
@@ -204,8 +208,8 @@ class CoreImportHandler(object):
 
         for module_name in module_names:
 
-            # don't re-import this module. we always use the first one added
-            # to `sys.meta_path` anyway.
+            # just to be safe, don't re-import this module.
+            # we always use the first one added to `sys.meta_path` anyway.
             if module_name == __name__:
                 continue
 
@@ -214,29 +218,12 @@ class CoreImportHandler(object):
 
             if pkg_name in old_namespaces and pkg_name in self._namespaces:
                 # the package name exists in an old core namespace and in the
-                # new core namespace. we need to re-import it, but first,
-                # delete it from sys.modules so that the custom import can run.
-                # This code leaves imported modules in place for use in already
-                # running code for scenarios where an old namespace doesn't
-                # exist in the new core (probably very rare).
-                module_names_to_import.append(module_name)
+                # new core namespace. we delete it from sys.modules so that
+                # the custom import can run.
                 del sys.modules[module_name]
 
-        """
-        # now go through the list of previously imported modules and re-import
-        # them. these imports will run through the custom `find_module` and
-        # `load_module` and be imported from the newly set core path.
-        for module_name in module_names_to_import:
-            try:
-                __import__(module_name)
-            except ImportError:
-                # The existing module could not be re-imported. It may not be
-                # necessary with the new core, so just ignore it.
-                # Future attempts to import or use this module may raise, but
-                # those exceptions will likely be more useful if raised in
-                # context.
-                pass
-        """
+        # clear out the previously found module info
+        self._module_info = {}
 
         # release the lock so that other threads can continue importing from
         # the new core location.
@@ -257,3 +244,40 @@ class CoreImportHandler(object):
         :returns: a list where each item is a namespace str
         """
         return self._namespaces
+
+def swap_core(core_path):
+    """Swap the current core with the core located at the supplied path
+
+    Actually just unloads the existing core and ensures an import handler
+    exists that points to the supplied core path. When this method completes,
+    all core namespaces will be removed from `sys.modules`. The required
+    modules will need to be reimported after this method is called.
+
+    :param core_path: The path to the new core to use upon import.
+
+    """
+
+    # see if there's already a core import handler in use
+    import_handler = None
+    for handler in sys.meta_path:
+        if isinstance(handler, CoreImportHandler):
+            import_handler = handler
+            break
+
+    if not import_handler:
+        # no import handler yet, create a new one
+        import_handler = CoreImportHandler(core_path)
+
+        # add the new import handler to the meta path so that it starts
+        # taking over core-related imports
+        sys.meta_path.append(import_handler)
+
+    if core_path != import_handler.core_path:
+        # the core we want to load differs from the one the import handler
+        # is using. we'll set it so that future imports of core namespaces
+        # use this new location
+        import_handler.set_core_path(core_path)
+
+        # NOTE: once this is called, any imported modules in the core python
+        # namespaces will be unloaded. the calling code will need to reimport
+        # everything necessary before continuing.
